Marked fields with `[NotNull]` and `[DisallowNull]`
There isn't a constructor for initializing the values of fields like e.g. `Key` and, therefore, they must be marked as nullable. In practice, however, we will never create these classes without those values being immediately provided afterwards, and so we can mark them as `[NotNull]` to prevent downstream code from needing to be unnecessarily defensive. Likewise, to ensure that anyone attempting to set them with a null value gets a warning, we add `[DisallowNull]`—though, in practice, this likely won't happen since these objects are expected to generally be created by either the `TopicExtensions.Export()` method or a to-be-developed deserialization process.