Established test cases for (de)serialization
While (de)serialization should "just work", it's important to test the process explicitly since a) this is the key use case for evenhaving the `OnTopic.Data.Transfer` objects in the first place, and b) should we choose to customize the (de)serialization (which was the original goal), this will provide a solid foundation to work off of. This also helped expose some issues—which will be corrected shortly—such as needing `setters` on collections for them to be deserialized.