Introduced `IncludeExternalReferences`
Added an `IncludeExternalReferences` property to `ExportOptions`, set the default value to `false`, and added unit tests to evaluate these scenarios. Be aware that this changes the previous behavior, which included all references regardless of whether or not they were within the `CurrentScope` of the `Export()`. That said, since we haven't deployed this yet, that shouldn't be a concern.