Establish LastModifiedImportStrategy
The `LastModifiedImportStrategy` enum defines a number of predetermined strategies for handling the `LastModified` and `LastModifiedBy` attributes. These require special handling as they are normally calculated fields, and it may not be appropriate to inherit the values from external topic graphs—and especially if those fields are exposed as part of a public template (e.g., to denote authorship).