Introduce logic for tracking changes
Ideally, the `LastModified` and `LastModifiedBy` records should only be applied if the records have been changed. Unfortunately, we don't have a quick and easy way of determing if the relationships have been changed. As a result, we have to go through somewhat convoluted logic to manually compare the before and after state.