Only set relationships and `LastModified(By)` if `IsDirty
Use the new `IsDirty` tracking (see #4340b6f) to determine if the relationships, `LastModified`, and `LastModifiedBy` attributes should be updated. The relationships bit is a bit of a minor optimization. Since we're going through the logic of determining if the relationships differ, we can avoid merging the relationship trees. That may help cut down on the overhead of checking `IsDirty` to begin with. (An alternative would have been to make the relationship merging logic more complicated so that it could intelligently detect changes.)