Establish tests for new `LastModified` handling
These tests confirm that the `Inherit`, `Current`, `System`, and `TargetValue` strategies all work as expected for the `LastModified` attribute. These are similar to the tests established in #79431e9, except that they evaluate the `LastModified` date as opposed to the `LastModifiedBy` user, since these can be set separately, and have (slightly) different business logic.