Updated `IsDirty` tests to use new arrange helper
The new `GetTopicWithNewerTopicData()` helper function (#0174e65, #79db4d9) has been further extended to additionally accept a(n optional) `sourceRelationshipKey` and `isDirty` parameter so that it can be used with the `*_TopicDataWith(Out)Changes_*` unit tests. This greatly reduces the amount of code, and helps centralize updates to how the unit tests are arranged.