Added missing unit tests
There are two scenarios which don't appear to be working, and which expose holes in our unit tests. Given this, the `Import_TopicDataWithChild_MapsExistingTopic` was modified to evaluate mapped properties on children, and not just the count of children. Further, a new `ImportAsReplace_TopicDataWithNestedTopic_DeletedOrphanedChildren` test was introduced, which checks the count on nested topics specifically, and also ensures referential integrity. Both of these tests will currently fail.