Evaluate `DeleteUnmatched*` based on parent, not child
The logic for deleting unmatched children (`DeleteUnmatchedChildren`) and nested topics (`DeleteUnmatchedNestedTopics`) incorrectly evaluated the state of the _child_, not the _parent_. As a result, it would, for example, delete entirely nested topic containers, including _all_ children, if `DeleteUnmatchedNestedTopics` was selected, instead of waiting until it had recured down to the nested topic container and then evaluate its children. This would also create entity integrity issues as it would effectively delete and replace the existing entity, wiping out its entire history, and giving it a new `Id`.