Remove unnecessary (and broken) `DeleteUnmatchedNestedTopics` logic
Not only will this logic not work—since it's trying to delete an item in a list its enumerating over—but it's also unnecessary, but even if it did work, it would be incorrectly deleting all nested topic _containers_, and not simply unmatched nested topics. See #0c5eb98 for a similar bug with the previous code block. Finally, this code is redundant in purpose with that previous block—whch, now that it's fixed, makes the entire _intent_ of this code block unnecessary.