Introduced unit test for reserved attributes
There is a classification of reserved attributes which are not intended to be exported as attributes. These include core elements such as `Key`, `ParentID`, `ContentType`, and `TopicID` which are handled on `TopicData`. The `ParentID` and `TopicID` are especially important as the topic identifiers aren't expected to translate as part of data transfers, and so these need to be converted into fully-qualified paths for the sake of exporting.

This functionality was already in place, but wasn't working correctly, particularly due to differences in case sensitivity. Adding a unit test both to evaluate this functionality, but also to test the case sensitivity.