Made reserved attribute lookup case insensitive
The previous commit (e2b5bae) fixed a bug with the reserved attribute keys not using the correct case. This should fix the bug in most cases. But, in concept, attribute keys _shouldn't_ be case sensitive. They certainly aren't treated as such by e.g. `AttributeValueCollection.GetValue()`, for instance. As such, while the casing should be consistent in our code, an even better long-term solution is to make sure that the lookup itself is not case sensitive.

This commit ensures that the lookup of reserved keywords is not case sensentive.