Incorporate new `GetByUniqueKey()` method
A recent update to the **OnTopic Library** includes a new `GetByUniqueKey()` extension method on `Topic` which allows a topic to be retrieved by its unique key from anywhere within the current topic graph. That eliminates the need for both a) finding the `rootTopic`, and b) defining a custom `FindUniqueKey()` function, as was previously done by the `TopicExtensions` class.