Updated to OnTopic Library preview
This preview release of the **OnTopic Library** includes new features such as the `GetByUniqueKey()` extension that can be used to replace helper functions in `TopicExtensions` (930538e).

Obviously, this will need to be updated to the release version once it's available—and prior to this update to the **OnTopic Data Transfer** library being released.