Remove `isDirty` check for relationships
Previously, the `isDirty` check would determine if any relationships had changed and, if so, include the byline (`LastModifiedBy`) and dateline (`LastModified`). This ensured that the byline and dateline weren't included if nothing had changed—neither attributes nor relationships.

While this was a nice feature, a similar feature has since been introduced into the forthcoming version of the **OnTopic Library** in which the byline and dateline are excluded from `Save()` if no attributes have changed. This isn't as sophisticated in that it doesn't include the relationships, and can thus lead to false negatives if the only change is to the relationships.

Because of this feature, however, including the relationships in this check is irrelevant, as the downstream `TopicRepositoryBase` and e.g. `SqlTopicRepository` won't write the byline and dateline if none of the attributes have changed. As such, factoring the relationships into this will, at best, introduce confusion (e.g., by updating the in-memory state, but not updating the underlying repository store).

In addition, since we had already been validating whether or not relationships had changed, we were using this to conditionally determine whether the relationships should be imported. That was a minor performance optimization since we already knew whether that was necessary. We could keep the existing logic and _just_ use it for that condition, but the cost of prevalidating if any relationships have changed is likely greater than just updating the relationships, plus requires more code, so we're just removing all of it.