Translate `TopicId` to `UniqueKey` in attributes on `Export()`
Some types of attributes—such as `TokenizedTopicList`, `TopicList`, and `TopicReference`—create pointers to other topics. These can be stored in a variety of formats, including topic key, unique key, or topic ID. If the latter, that can introduce problems when exporting the data, as those IDs won't be maintained when importing them into a new database.

To mitigate this, a new `GetUniqueKey()` helper function and corresponding `getAttributeValue()` local function is introduced into the `Export()`  extension method. If an attribute ends with `ID`, is a numeric numeric value, and maps to a topic with the same `TopicID` in the topic graph, then the value is replaced with the topic's `GetUniqueKey()` on export.

There is a risk of false positives here. For instance, GoldSim has an attribute named `PageID` which is used for a legacy page identifier, and doesn't correspond to topics. Yet, it's conceivable that the values will overlap with valid `TopicID` values, thus getting replaced. Similarly, if there is an attribute named e.g. `OpeningBid` which has an integer value, that could also trigger a false positive.

At this point, there's no way to mitigate these false positives. We may add options later to better control the cases where these are exported to help prevent these. For now, however, this provides the initial implementation.

Includes unit test.