Introduce `ExportOptions.TranslateTopicPointers`
In a recent feature improvement, we added the ability to translate topic pointers from `Topic.Id` to `Topic.GetUniqueKey()` on `Export()` and then back to `Topic.Id` on `Import()`. This resolves an issue where attribute types such as `TokenizedTopicList` may write topic pointers as attributes which then point to a different topic when imported into a different database.

While this resolves a major issue, it also introduces room for false positives. Namely, any attribute which ends with `Id` and have a numeric value will be treated as topic pointers. So, for example, GoldSim has a `PageID` attribute which maps to their legacy database identifiers; we wouldn't want _those_ to be treated as topic pointers.

To provide more control over this, the `ExportOptions.TranslateTopicPointers` option was introduced. This disables the translation of topic pointers by default, while optionally allowing them to be enabled.

In addition, as part of this, I've updated the logic to honor the preexisting `IncludeExternalReferences`. If a topic pointer falls outside the `ExportOptions.ExportScope` and `IncludeExternalReferences` isn't enabled, then the attribute is skipped. Similarly, if the topic pointer references an invalid topic, then the attribute is also removed. This mitigates a scenario where an orphaned or invalid value is maintained—but ends up pointing to an actual topic when imported into a new database.

In addition to integrating this with `Topic.Export()`, I've also moved the unit tests over to `ExportOptionsTest` and added new tests for evaluating more nuanced cases. For instance,