Prevent `Export()` of `null` or empty attributes
Null or empty attributes aren't added to the `AttributeValue` collection on `Load()`. With the new logic for validating topic pointers (b6a4e63), however, `GetUniqueKey()` will return `null` if the topic pointer is either invalid or outside the scope of the export. In these cases, we don't want the `null` value being `Export()`ed, as this would lead to overwriting any preexisting values being deleted in `Import()`.

To mitigate that, an explicit check is introduces to ensure that the attribute is _only_ added to the `TopicData.Attributes` collection if the value is not `null` or empty.