Modified `TranslateTopicPointers` default to `true`
In most cases, the logical default for `TranslateTopicPointers` (b6a4e63) should be `true`. This is because out-of-the-box attribute types (such as `TopicReference`, `TopicList`, and `TokenizedTopicList`) can use it to save topic pointers as attributes.

As such, I've modified `TranslateTopicPointers` to be opt-out instead of opt-in. This way, it can still be disabled in case there is a serious risk of false positives—_and_ that risk outweighs the risk of false negatives.

If this becomes a serious issue, we can reevaluate providing more granular control—e.g., supporting a black list of invalid topic pointer keys which the caller can supply. But that's an edge case. A best practice remains to only end topic pointers with `Id`, and there are very rare cases where this is violated (e.g., in GoldSim's legacy `PageID`).

Since this is now the default behavior of `Export()`, I've moved the unit tests related to this functionality _back_ to `TopicExtensionsTest` from `ExportOptionsTest`. (This entire distinction is somewhat debatable given how tightly integrated they are, but if we want to change that there's a lot more refactoring of the tests we need to provide.)