Resolved bug in recursively resolving orphaned relationships
In a previous feature (f358ef3), I added the ability to resolve orphaned relationships. This failed outside of direct hierarchical relationships, however, due to the `unresolvedRelationships` collection not being properly relayed upon recursion. This meant that when a particular branch finished importing, any descendents would be picked up. But those unresolved references would be lost for sibling branches.

This issue is fixed, and the unit test is updated to detect these scenarios in the future.