Migrate GitVersion to `ContinuousDeployment`
By default, GitVersion only uses `ContinuousDeployment` on the `develop` branch, and `ContinuousDelivery` on all other branches. This creates a problem when attempting to publish preview releases of particular for integration testing, as that means the versions need to be manually bumped using a tag. By setting the default to `ContinuousDeployment`, the build version will increment with every commit, thus ensuring unique versions. The `master` branch is kept on `ContinuousDelivery` so that it requires manual incrementing via tags.