Ensure a `None` option for enums
Technically, we never expect a user to select `None` for either `ImportStrategy` or `LastModifiedImportStrategy`. That said, as the default value for an enum is 0, it's useful to explicitly define the `None` option to avoid potential bugs or unexpected behavior. This makes this scenario more explicit, should it arise. This satisfies CA1008.