Establish a new `AssemblyInfo` for `OnTopic.Data.Transfer`
These used to be created by default with new projects. They no longer are since most properties have been moved to the `csproj` file. There are some properties that are still useful to set on the assembly, however. Most notably, this sets the library to CLS compliant, which will allow downstream subscribers to trust that it adheres to CLS conventions. While I was at it, I made any internal members visible to the test assembly, and explicitly set the GUID from the solution (`.sln`) file.