Ensured culture invariant timestamp conversion in unit tests (fixed)
In a previous update, I applied culture invariant string conversion for the dates (30c8677). In doing so, however, I inadvertantly introduced a bug since the GetTopicWithNewerTopicData()` helper function wasn't performing the same conversion. Ultimately, since these dates are being set and read in the same class, it doesn't actually matter which culture they use, as it will always be internally consistent. Keeping it explicit, though, helps avoid any ambiguity.

(Technically, we should probably do this in the editor itself when setting the `LastModified` attribute, but that's a separate task that will be handled later.)