Updated reference to `AttributeValue` to new `AttributeRecord`
In OnTopic 5.0.0, the name of the `AttributeValue` type is changed to `AttributeRecord` as part of a broader effort to unify the naming conventions for combining both data (i.e., `Key`, `Value`) and metadata (i.e., `IsDirty`, `LastModified`). This is also used for the new `TopicReferenceRecord` type. Further, it mitigates the awkward `AttributeValue.Value` syntax.