Fixed attribute check for `TopicID` in unit tests
In OnTopic 5.0.0, the `BaseTopic` (previously `DerivedTopic`) is no longer stored in the `Topic.Attributes` collection as a `TopicID` attribute. Instead, it is now stored in the strongly typed `Topic.References` collection. As such, checking the `TopicID` is no longer expected to be a valid check. However, the existing equality check against `baseTopic` continues to work.

While I was at it, I renamed the key for base topics from `Derived` to `Base`. This doesn't impact the tests, but is more intuitive as a name for human readers of the test or its debugging data.