Introduced a test for evaluating the legacy `DerivedTopicKey`
In OnTopic 5.0.0, the `Topic.DerivedTopic` was renamed to `Topic.BaseTopic` (55e952b). To maintain consistency, the `TopicData.DerivedTopicKey` was renamed to `TopicData.BaseTopicKey` (22b9866). But previously exported data will still have the legacy `DerivedTopicKey` property set in the JSON. This test determines whether that is correctly imported and associated with the new `BaseTopicKey` property.

Note: This test will fail as of this commit, as there's currently no effort to handle this special requirement. Subsequent commits will resolve this issue.