Allow `BaseTopicKey` to be defined based on `DerivedTopicKey`
In a previous commit, the `DerivedTopicKey` property was renamed to `BaseTopicKey` to maintain parity with OnTopic 5.0.0's change of `Topic.DerivedTopic` to `Topic.BaseTopic`. Unfortunately, however, that prevents previously exported JSON files from being properly imported.

To mitigate that, I've reintroduced a (deprecated) `DerivedTopicKey` property, and set `BaseTopicKey`'s getter to fallback to it if `BaseTopicKey` isn't otherwise defined.

This has the unfortunate side-effect of writing a null `DerivedTopicKey` property to the exported JSON. This can be mitigated by including the `IgnoreNullValues` on the `JsonSerializerOptions`, which we recommend doing, and will later add in to the OnTopic Editor's integration.

Later, when we upgrade to .NET 5.x, we'll be able to handle this better by treating `DerivedTopicKey` as a private field with the `[JsonInclude]` attribute along with the `[JsonIgnore(JsonIgnoreCondition.WhenWritingDefault)]` attribute. But, for now, we want to maintain backward compatibility with .NET Core 3.x, and so those aren't yet available.