Introduced `References` collection on `TopicData`
This can reuse the existing `AttributeDataCollection` and `AttributeData` item type since topic references get serializes with a string `Value` and are tracked via a `LastModified` value.

In the future, we may want to update the name of the `AttributeDataCollection` and `AttributeData` classes to make them more generic, but we'll address that separately. (It doesn't impact either the serialization or the deserialization directly.)