Generalized handling of unresolved associations
Previously, the (local) `unresolvedRelationships` collection handled tracking of relationships that could not be resolved on the initial pass, and would attempt to wire them up in a subsequent pass. This has now been generalized to support both `Topic.Relationships` and `Topic.References`. To acknowledge that, the variable has been renamed to `unresolvedAssociations`.