Ensure `BaseTopic` can be imported via `TopicData.References`
In OnTopic 5.0.0, the `BaseTopic` (previously `DerivedTopic`) is now handled via the (new) `Topic.References` feature. This prevents the need to specially account for it, as there's now first-class support for this _type_ of association, independent of the specifics of `BaseTopic`.

With topic references now accounted for during the `Import()` process (6d5938b), there's also no need for specialty handling of `BaseTopic`, so long as it's passed via a `TopicData.References`. This test sets up an `AttributeData` with a key of `BaseTopic`, adds it to the `TopicData.References` collection, and ensures that it is successfully wired up.