Fixed alignment of `BaseTopicKey` in unit tests
With the rename of `DerivedTopicKey` to `BaseTopicKey`, the alignment of the variables was off.

While I was at it, I also updated one of the assignments to use the root value from the `TopicData` object, so as to not repeat string values. (This is consistent with how we handle this situation in other tests.)