Rely on `TopicData.References` to handle `BaseTopic`
For new exports, there's no need to set the `BaseTopicKey`; this will be handled automatically by the `TopicData.References`, since `Topic.BaseTopic` corresponds to a `Topic.References` entry.

In fact, really, we only need the `BaseTopicKey` for legacy support—which calls into question the entire basis for introducing the `BaseTopicKey` to begin with (4705b29). We'll likely address that in a subsequent update.

As part of this, I also updated the unit test to look for the `BaseTopic` inside of the `TopicData.References` collection, instead of within the `BaseTopicKey`.