Unify `BaseTopicKey` handling with `TopicData.References`
Since `TopicData.References` is now the preferred way of handling `Topic.BaseTopic` (7c20dd0), the handling of `BaseTopicKey` is really only present for backward compatibility. Given that, instead of separately handling the `BaseTopicKey`, we should just use it to establish a `BaseTopic` entry in `TopicData.References`, if one doesn't already exist.

This has a few benefits. First, it ensures that the `ImportOptions` are honored, such as honoring `ImportStrategy.Add`, if set, instead of always overwriting the `BaseTopic`—and, therefore, the associated entry in `Topic.References`. Second, it better accounts for unexpected scenarios where both the `BaseTopicKey` and a `BaseTopic` reference are set (in that case, the newer `BaseTopic` reference should always be preferred). Third, it centralizes the lookup of referenced topics to one location.