Update `Import()` tests for `BaseTopic` to use `TopicData.References`
Most of the `Import()` tests for evaluating `BaseTopic` continued to set the `TopicData.BaseTopicKey`, instead of adding an `AttributeData` record to `TopicData.References`, as is now preferred (7c20dd0). With the exception of the sole test for evaluating backward compatibility (f1ddafe), these tests have been updated to use the new, preferred format.

As part of this, the unit test names and descriptions have been updated to better reflect this functionality.