Update Remove `BaseTopicKey` from deserialization tests
Now that `BaseTopicKey` is no longer set as part of `Export()`—and is therefore not expected from any OnTopic Data Transfer 3.0.0 library exports—the deserialization test should no longer reference `BaseTopicKey`.

Note: The exception to this is for unit tests evaluating backward compatibility with the legacy JSON format—but, in that case, they should instead reference `DerivedTopicKey` (18483b8).