Remove unnecessary `BaseTopicKey`
In a previous feature, we introduced the `BaseTopicKey` as a replacement for the legacy `DerivedTopicKey` (4705b29). This was useful for maintaining backward compatibility while maintaining consistency with the new `Topic.BaseTopic` naming convention (55e952b).

Now that we've fully implemented the new `Topic.References` capabilities into the `Export()` (f156c25) and `Import()` (6d5938b) methods, however, and rely primarily on those for handling `BaseTopic` references (7c20dd0, 76a506d), there's no need to maintain the `BaseTopicKey` property. In fact, doing so only introduces confusion in terms of how it relates to the `TopicData.References` version of `BaseTopic` (1db08fb).

This effectively reverts most of the functionality introduced as part of the `feature/BaseTopicKey` branch (4705b29), with the exception that `DerivedTopicKey` continues to be marked as deprecated via the `[Obsolete()]` attribute.