Import legacy topic references as topic references
If legacy topic references exist—i.e., attributes whose `Key` ends in `Id` and who point to a `UniqueKey()`—then strip the `Id` extension and import them into the new `TopicData.References` collection. Do not do this if a topic reference with the new `Key` already exists.

If an existing `Key` is found, or the value doesn't appear to be a `UniqueKey()`, then continue to import it into `TopicData.Attributes`. This is a change in behavior from the previous functionality, which would fail to import it at all if the value could not be translated.

This change has three benefits.

First, it unifies the legacy topic references with the new topic references. As this occurs automatically as part of the OnTopic 5.0.0 migration script, our assumption is that any JSON with these legacy topic references is from before that migration and, therefore, should be updated in the same way that the e.g. SQL data was updated.

Second, it ensures that `unresolvedAssociations` are properly tracked, such that topic references that haven't yet been established (e.g., because they exist later in the topic graph that's being imported) will be picked up after `Import()` has finished importing the entire topic graph.

Finally, it helps standardize the code for handling topic references, instead of having a separate process for handling legacy topic references and new topic references.

This also allows us to get rid of the legacy `getAttributeValue()` and `GetTopicId()` helper functions, which were needed exclusively for the support of the legacy topic reference handling. On the net, this simplifies the code.